fin confirm before finish

+++ www/src/views/basic/dict/list.js
@@ -57,15 +57,20 @@ export default {
       })
     },
     remove(id, index) {
-      deleteModel(this.KEY, id).then(({data})=> {
-        this.list.splice(index, 1)
-        this.$message({
-          showClose: true,
-          message: '删除成功',
-          type: 'success'
+      this.$confirm(`确认删除: ${this.list[index].label}?`,  {
+        type: 'warning'
+      }).then(() => {
+        deleteModel(this.KEY, id).then(({data})=> {
+          this.list.splice(index, 1)
+          this.$message({
+            showClose: true,
+            message: '删除成功',
+            type: 'success'
+          })
+          this.isSynced = false
         })
-        this.isSynced = false
-      })
+      }).catch(() => {})
+      
     },
     removeSubItem(index) {
       this.currRow.hasChanged = true
