+++ www/src/views/router-menu/router/list.js
@@ -107,15 +107,20 @@ export default {
       })
     },
     remove(id, index) {
-      deleteModel(this.KEY, id).then(({data})=> {
-        this.list.splice(index, 1)
-        this.$message({
-          showClose: true,
-          message: '删除成功',
-          type: 'success'
+      this.$confirm(`确认删除?`,  {
+        type: 'warning'
+      }).then(() => {
+        deleteModel(this.KEY, id).then(({data})=> {
+          this.list.splice(index, 1)
+          this.$message({
+            showClose: true,
+            message: '删除成功',
+            type: 'success'
+          })
+          this.isSynced = false
         })
-        this.isSynced = false
-      })
+      }).catch(() => {})
+      
     },
     getEntity(entityId) {
       return [
